backend/requirements.txt
langchain
langchain-community
langgraph
langchain-groq
python-dotenv
torch
transformers
openai-whisper
audioread
pydub
--only-binary=all webrtcvad
ffmpeg-python
moviepy
tqdm
google-api-python-client
google-generativeai>=0.8.0
numpy>=1.20.0
soundfile>=0.12.1
SpeechRecognition>=3.10.0
librosa>=0.10.0
pyaudio>=0.2.13
python-dateutil>=2.8.2
scipy>=1.10.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
sounddevice>=0.4.6
webrtcvad>=2.0.10